% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACDC.prior.R
\name{ACDC.prior}
\alias{ACDC.prior}
\title{ACDC.prior function from Geiger}
\usage{
ACDC.prior(phy, decrease.max = 0.01, increase.max = 100)
}
\description{
ACDC.prior function from Geiger
}
\examples{
ACDC.prior()
}
