% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicc.R
\name{aicc}
\alias{aicc}
\title{aicc function}
\usage{
aicc(x, k, n)
}
\arguments{
\item{x}{likelihood}

\item{k}{parameters}

\item{n}{sample size}
}
\value{
aicc
}
\description{
aicc function
}
\examples{
aicc()
}
