% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchingTimesFossil.R
\name{BranchingTimesFossil}
\alias{BranchingTimesFossil}
\title{BranchingTimesFossil function from Graham Slater}
\usage{
BranchingTimesFossil(phy, tol = .Machine$double.eps^0.5)
}
\value{
list phylogeny and matrix of parameters
}
\description{
BranchingTimesFossil function from Graham Slater
}
\examples{
phylogMean()
}
