% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladeMode.R, R/cladeModeSim.R
\name{cladeMode}
\alias{cladeMode}
\alias{cladeMode}
\title{Analyse nested modes of evolution}
\usage{
cladeMode(n = 1, phy, node, model = c("BM", "EB", "nestedEB",
  "nestedEBRate", "rateShift", "nestedOU", "OU"), beta = 1, a = NULL,
  alpha = 0.05, shiftRate = 0.5, ebBeta = 2)

cladeMode(n = 1, phy, node, model = c("BM", "EB", "nestedEB",
  "nestedEBRate", "rateShift", "nestedOU", "OU"), beta = 1, a = NULL,
  alpha = 0.05, shiftRate = 0.5, ebBeta = 2)
}
\arguments{
\item{n}{number of datasets to simulate (default = 1)}

\item{phy}{tree in APE formate}

\item{node}{number of node at which to estimate nested evolutionary mode. If FALSE (default) the model searches across all nodes}

\item{beta}{the value of the Brownian motion rate (default = 1)}

\item{a}{the value of the early burst parameter. If NULL the value is the midpoint of the expected early burst values}

\item{shiftRate}{the value of the shiftRate parameter (default = 1)}

\item{ebBeta}{the value of the scaler for the nested early burst model (default = 2)}

\item{y}{named vector of trait data}

\item{minSize}{the minimum size of node at which to select a nested rate}

\item{mode}{of evolution for the model. One of "BM", "EB", "nestedEB", "nestedEBRate", "rateShift", "nestedOu", "rateSlow")}

\item{ncores}{number of seperate cores to simultaneously estimate rate shifts - defaults to 1}

\item{contrasts}{logical (default = FALSE) as to whether to estimate parameters with likelihood search or independent contrasts}

\item{returnPhy}{logical (default = FALSE) return re-scaled phylogeny}

\item{phy}{tree in APE formate}

\item{y}{named vector of trait data}

\item{node}{number of node at which to estimate nested evolutionary mode}

\item{mode}{of evolution for the model. One of "BM", "EB", "nestedEB", "nestedEBRate", "rateShift", "nestedOu", "rateSlow")}

\item{mode}{of evolution for the model. One of "BM", "EB", "nestedEB", "nestedEBRate", "rateShift", "nestedOu", "rateSlow")}
}
\value{
results a list containing the model parameters, log likelihood, aicc, optionally the rescaled tree

y a matrix of simulated values
}
\description{
Analyses various nested modes of trait evolution

Simulate various nested modes of trait evolution
}
\examples{
cladeMode()
cladeMode()
}

