% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladeModeNode.R
\name{cladeModeNode}
\alias{cladeModeNode}
\title{Analyse nested modes of evolution - internal function}
\usage{
cladeModeNode(phy, y, node, model = c("BM", "EB", "nestedEB", "nestedEBRate",
  "rateShift", "nestedOU"), cont, returnPhy, mErr)
}
\description{
Internal function to analyses various nested modes of trait evolution
}
\examples{

}
