% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladeModeSim.R
\name{cladeModeSim}
\alias{cladeModeSim}
\title{Simulate nested modes of evolution}
\usage{
cladeModeSim(n = 1, phy, node, model = c("BM", "EB", "nestedEB",
  "nestedEBRate", "rateShift", "nestedOU", "OU"), beta = 1, a = NULL,
  alpha = 0.05, shiftRate = 0.5, ebBeta = 2)
}
\arguments{
\item{n}{number of datasets to simulate (default = 1)}

\item{phy}{tree in APE formate}

\item{node}{number of node at which to estimate nested evolutionary mode}

\item{beta}{the value of the Brownian motion rate (default = 1)}

\item{a}{the value of the early burst parameter. If NULL the value is the midpoint of the expected early burst values}

\item{shiftRate}{the value of the shiftRate parameter (default = 1)}

\item{ebBeta}{the value of the scaler for the nested early burst model (default = 2)}

\item{y}{named vector of trait data}

\item{mode}{of evolution for the model. One of "BM", "EB", "nestedEB", "nestedEBRate", "rateShift", "nestedOu", "rateSlow")}

\item{mode}{of evolution for the model. One of "BM", "EB", "nestedEB", "nestedEBRate", "rateShift", "nestedOu", "rateSlow")}
}
\value{
y a matrix of simulated values
}
\description{
Simulate various nested modes of trait evolution
}
\examples{
cladeModeSim()
}
