% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebSubClade.R
\name{ebSubClade}
\alias{ebSubClade}
\title{Rescale phylogeny with early burst values}
\usage{
ebSubClade(phy_tree, a, node, branches, times, originTime)
}
\arguments{
\item{phy_tree}{input phylogeny}

\item{a}{}

\item{originTime}{}
}
\value{
rescaled vcv matrix
}
\description{
Internal function to rescale phylogeny with early burst values
}
\examples{
cladeMode()
}

