% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebSubClade_rescale.R
\name{ebSubClade_rescale}
\alias{ebSubClade_rescale}
\title{Rescale phylogeny with early burst values}
\usage{
ebSubClade_rescale(phyIn, aInput, node_in)
}
\arguments{
\item{phyIn}{input phylogeny}

\item{a}{Input}
}
\value{
rescaled vcv matrix
}
\description{
Internal function to rescale phylogeny with early burst values
}
\examples{
cladeMode()
}

