% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNodes.R
\name{findNodes}
\alias{findNodes}
\title{findNodes function}
\usage{
findNodes(phylo, minSize = 0.1)
}
\arguments{
\item{minSize}{(default = 0.1)}
}
\value{
node
}
\description{
findNodes function
}
\examples{
findNodes()
}
