% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodes.R
\name{getNodes}
\alias{getNodes}
\title{getNodes function}
\usage{
getNodes(ttr)
}
\arguments{
\item{x}{likelihood}

\item{k}{parameters}

\item{n}{sample size}
}
\value{
aicc
}
\description{
getNodes function
}
\examples{
getNodes()
}

