% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unit_tree.phylo.R
\name{make_unit_tree.phylo}
\alias{make_unit_tree.phylo}
\title{make_unit_tree.phylo function from arbutus}
\usage{
\method{make_unit_tree}{phylo}(x, data, ...)
}
\description{
make_unit_tree.phylo function from arbutus
}
\examples{
make_unit_tree.phylo()
}
