% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_sim.R
\name{model_sim}
\alias{model_sim}
\title{model_sim}
\usage{
model_sim(phy, sim_data, n_core = 1, nReps = 1000)
}
\arguments{
\item{phy}{input phylogeny}

\item{n_core}{default=1}
}
\description{
summarise simulations
}
\examples{
model_sim()
}
