% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoClade.R
\name{monoClade}
\alias{monoClade}
\title{identify nodes of a monophyletic clade}
\usage{
monoClade(phylo, node)
}
\arguments{
\item{phylo}{dated tree in ape format}
}
\value{
storeNodes
}
\description{
identify nodes of a monophyletic clade
}
\examples{
monoClade()
}
