% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoTips.R
\name{monoTips}
\alias{monoTips}
\title{monoTips}
\usage{
monoTips(phylo, node)
}
\arguments{
\item{phylo}{input phylogeny}

\item{node}{}
}
\description{
monoTips calculates the proportion of the tree diversity represented by a clade
}
\examples{
monoTips()
}
