% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeDes.R
\name{nodeDes}
\alias{nodeDes}
\title{nodeDes}
\usage{
nodeDes(phylo, node)
}
\arguments{
\item{phylo}{input tree in APE formate}

\item{node}{}
}
\value{
a vector of tip labels
}
\description{
identifies the tip descendents of an ancestral node
}
\examples{
library(geiger)
set.seed(30)
# simulate tree with 20 species
simTree <- rcoal(20)
nodeDes(simTree, 22)
}
