% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouMatrix.R
\name{ouMatrix}
\alias{ouMatrix}
\title{ouMatrix function from Graham Slater}
\usage{
ouMatrix(vcvMatrix, alpha)
}
\value{
list phylogeny and matrix of parameters
}
\description{
ouMatrix function from Graham Slater
}
\examples{
ouMatrix()
}
