% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogMean.R
\name{phylogMean}
\alias{phylogMean}
\title{phylogMean function from Graham Slater}
\usage{
phylogMean(phyvcv, data)
}
\value{
list phylogeny and matrix of parameters
}
\description{
phylogMean function from Graham Slater
}
\examples{
phylogMean()
}
