% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEB.R
\name{plotEB}
\alias{plotEB}
\title{plot EB phylogeny}
\usage{
plotEB(phyIn, aInput, node_in, rate)
}
\arguments{
\item{phyIn}{input phylogeny}

\item{a}{Input}
}
\value{
rescaled vcv matrix
}
\description{
plot EB phylogeny
}
\examples{
cladeMode()
}

