% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomNode.R
\name{randomNode}
\alias{randomNode}
\title{randomNode}
\usage{
randomNode(phylo, size = 0.25, overSize = TRUE)
}
\arguments{
\item{phylo}{input phylogeny}

\item{size}{minimum n tips of descending node (as given by a proportion of total tips)}

\item{overSize}{return node ancestral to n tips as specified by 'size' or not (default = TRUE)}
}
\description{
randomNode
}
\examples{
nodeDes()
}
