% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateClade.R
\name{rateClade}
\alias{rateClade}
\title{rateClade function}
\usage{
rateClade(phy_tree, branches, rate = 1)
}
\arguments{
\item{rate}{default = 1}
}
\value{
phy_tree
}
\description{
rateClade function
}
\examples{
rateClade()
}
