% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reScalePhylo.R
\name{reScalePhylo}
\alias{reScalePhylo}
\title{reScalePhylo function}
\usage{
reScalePhylo(phy, y, param, model = c("BM", "EB", "nestedEB", "nestedEBRate",
  "rateShift", "rateSlow", "nestedOU"))
}
\value{
list phylogeny and matrix of parameters
}
\description{
reScalePhylo function
}
\examples{
reScalePhylo()
}

