% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalePIC.R
\name{rescalePIC}
\alias{rescalePIC}
\title{rescalePIC function}
\usage{
rescalePIC(phy, y)
}
\arguments{
\item{phy}{dated tree in ape format}

\item{y}{named vector of trait values}
}
\value{
phy a rescaled phylogeny
}
\description{
rescalePIC a function to produce a re-scaled phylogeny using IC
}
\examples{
rescalePIC()
}
