% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subClade.R
\name{subClade}
\alias{subClade}
\title{subClade function}
\usage{
subClade(phy_tree, a, node, branches, times, originTime, rate = 1)
}
\arguments{
\item{rate}{default = 1}
}
\value{
phy_tree
}
\description{
subClade function
}
\examples{
subClade()
}

