% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSimToApe.R
\name{treeSimToApe}
\alias{treeSimToApe}
\title{convert treeSim to APE}
\usage{
treeSimToApe(treeSimPhy)
}
\arguments{
\item{treeSimPhy}{phylogeny from treeSim}
}
\value{
trApe tree in APE format
}
\description{
convert treeSim to APE
}
\examples{
phyIC()
}
