% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiccWeight.R
\name{aiccWeight}
\alias{aiccWeight}
\title{aicc weight function}
\usage{
aiccWeight(aicc)
}
\arguments{
\item{aicc}{input aicc values}
}
\value{
aicc
}
\description{
aicc weight function
}
\examples{
aiccWeight()
}

