% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cladeMode.R
\name{cladeMode}
\alias{cladeMode}
\title{Analyse nested modes of evolution}
\usage{
cladeMode(phy, y, node = F, minSize = 5, mode = c("BM", "EB", "nestedEB",
  "nestedEBRate", "rateShift", "nestedOU", "rateSlow"), Ncores = 1,
  contrasts = F, returnPhy = FALSE)
}
\arguments{
\item{phy}{tree in APE formate}

\item{y}{named vector of trait data}

\item{node}{number of node at which to estimate nested evolutionary mode. If FALSE (default) the model searches across all nodes}

\item{minSize}{the minimum size of node at which to select a nested rate}

\item{mode}{of evolution for the model. One of "BM", "EB", "nestedEB", "nestedEBRate", "rateShift", "nestedOu", "rateSlow")}

\item{contrasts}{logical (default = FALSE) as to whether to estimate parameters with likelihood search or independent contrasts}

\item{returnPhy}{logical (default = FALSE) return re-scaled phylogeny}

\item{ncores}{number of seperate cores to simultaneously estimate rate shifts - defaults to 1}
}
\value{
results a list containing the model parameters, log likelihood, aicc, optionally the rescaled tree
}
\description{
Analyses various nested modes of trait evolution
}
\examples{
cladeMode()
}

