% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.vcv.node.R
\name{split.vcv.node}
\alias{split.vcv.node}
\title{split.vcv.node}
\usage{
\method{split}{vcv.node}(phy, node)
}
\arguments{
\item{phy}{input phylogeny}

\item{node}{}
}
\value{
vcvMatrix vcv matrix
}
\description{
split.vcv.node produces two VCV matrices that represent a monophyletic subclade and the remaining phylogeny respectively
}
\examples{
set.seed(30)
#simulate tree with 20 species
simTree <- rcoal(20)
split.vcv.node(simTree, 22)
}
